requirements.txt
Django==5.1.1
psycopg2==2.9.9
django-filter==24.3
redis==5.0.8
celery==5.4.0
djangorestframework==3.15.2

